flutter_backend_logapp/requirements.txt

Flask
flask-cors
psycopg2-binary
requests
